+++ benefits/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v1.6.0] benefits
 ### Releases
 - [NPM v1.6.0] https://www.npmjs.com/package/@wingspan/benefits/v/1.6.0 - benefits
+
+## 2023-10-23 01:23:39
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0 https://docs.wingspan.app/openapi/6470f38d65c260000c025474
+- Speakeasy CLI 1.104.0 (2.169.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.7.0] benefits
+### Releases
+- [NPM v1.7.0] https://www.npmjs.com/package/@wingspan/benefits/v/1.7.0 - benefits